Model.remove() is useful to delete all the documents matching given criteria.
Modified the removeManyPeople function to delete all the people whose name is within the variable nameToRemove, using Model.remove(). we Pass it to a query document with the name field set, and a callback.

Note: The Model.remove() doesn’t return the deleted document, but a JSON object containing the outcome of the operation, and the number of items affected.