completed the filemetadata exercise with Multer
HTML form have enctype="multipart/form-data" and name="upfile" properties
This tells multer which field on the request it should look for the files in.

multipart/form-data, which is primarily used for uploading files. Multer will not process any form which is not multipart (multipart/form-data).

When we submit a file, we gave: the file name, type, and size in bytes within the JSON response.